+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/BlockBehaviourMixin.java
@@ -1,0 +1,20 @@
+package insane96mcp.iguanatweaksreborn.mixin;
+
+import insane96mcp.iguanatweaksreborn.module.misc.Tweaks;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(BlockBehaviour.class)
+public class BlockBehaviourMixin {
+    @Inject(method = "tick", at = @At("HEAD"))
+    public void onBlockTick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random, CallbackInfo ci) {
+        Tweaks.onSpongeTick(state, level, pos, random);
+    }
+}